fix builds

+++ tasks/mpi/anikin_m_summ_of_different_symbols/src/ops_mpi.cpp
@@ -90,7 +90,7 @@ bool anikin_m_summ_of_different_symbols_mpi::SumDifSymMPIParallel::validation() 
 
 bool anikin_m_summ_of_different_symbols_mpi::SumDifSymMPIParallel::run() {
   internal_order_test();
-  int loc_size = 0;
+  size_t loc_size = 0;
   if (com.rank() == 0) {
     loc_size = ((strlen(input[0])) + com.size() - 1) / com.size();
   }
